generic paginator, pagination to stats and leaderboard
